+++ modules/Server/src/main/java/serverside/Bike.java
@@ -73,7 +73,7 @@ public class Bike {
         System.out.println(re.getTotal_Distance());
         Statement st = dbConnection.createStatement();
         st.executeUpdate("UPDATE person SET Bike = Bike + "
-                + (int)(new CarbonCalculator(2).bike(re.getCarType(), re.getTotal_Distance())) + " WHERE Email = '" + email + "'");
+                + new CarbonCalculator(2).bike(re.getCarType(), re.getTotal_Distance()) + " WHERE Email = '" + email + "'");
 
         st.close();
         dbConnection.close();
@@ -101,7 +101,7 @@ public class Bike {
         ResultSet rs = st.executeQuery(
                 "SELECT Bike FROM person WHERE Email = '" + email + "'");
         rs.next();
-        int distance = rs.getInt("Bike");
+        double distance = rs.getDouble("Bike");
 
         Resource re = new Resource();
         passToken(token, re);
